Move helper methods to apm.js
